chore: fix pnpm install
