+++ packages/database/src/json-types.ts
@@ -1,9 +1,5 @@
 /* eslint-disable import/no-relative-packages -- required for importing types */
 /* eslint-disable @typescript-eslint/no-namespace -- using namespaces is required for prisma-json-types-generator */
-import type {
-  TSurveyFollowUpAction,
-  TSurveyFollowUpTrigger,
-} from "@formbricks/database/types/survey-follow-up";
 import { type TActionClassNoCodeConfig } from "@formbricks/types/action-classes";
 import type { TOrganizationAccess } from "@formbricks/types/api-key";
 import { type TIntegrationConfig } from "@formbricks/types/integration";
@@ -24,6 +20,7 @@ import {
 } from "@formbricks/types/responses";
 import { type TBaseFilters } from "@formbricks/types/segment";
 import { type TSurveyBlocks } from "@formbricks/types/surveys/blocks";
+import type { TSurveyFollowUpAction, TSurveyFollowUpTrigger } from "@formbricks/types/surveys/follow-up";
 import {
   type TSurveyClosedMessage,
   type TSurveyEnding,
