+++ packages/types/package.json
@@ -20,7 +20,6 @@
     "node-html-parser": "7.1.0"
   },
   "devDependencies": {
-    "@formbricks/config-typescript": "workspace:*",
-    "@formbricks/database": "workspace:*"
+    "@formbricks/config-typescript": "workspace:*"
   }
 }
